+++ include/item_menu.h
@@ -64,7 +64,7 @@ struct BagMenuStruct
     const u8* contextMenuItemsPtr;
     u8 contextMenuItemsBuffer[4];
     u8 contextMenuNumItems;
-    u8 numItemStacks[POCKETS_COUNT];
+    u16 numItemStacks[POCKETS_COUNT];
     u8 numShownItems[6];
     s16 graphicsLoadState;
     u8 filler4[0xE];
